Fix failing test for BashWrapperBuilder

+++ src/test/groovy/nextflow/executor/BashWrapperBuilderTest.groovy
@@ -1827,14 +1827,17 @@ class BashWrapperBuilderTest extends Specification {
 
                 function nxf_mktemp() {
                     local base=\${1:-/tmp}
-                    [[ \$(uname) = Darwin ]] && mktemp -d \$base/nxf.XXXXXXXXXX || mktemp -d -t nxf.XXXXXXXXXX -p \$base
+                    if [[ \$(uname) = Darwin ]]; then mktemp -d \$base/nxf.XXXXXXXXXX
+                    else TMPDIR="\$base" mktemp -d -t nxf.XXXXXXXXXX
+                    fi
                 }
 
                 on_exit() {
                   exit_status=\${ret:=\$?}
                   printf \$exit_status > ${folder}/.exitcode
-                  rm -f "\$COUT" || true
-                  rm -f "\$CERR" || true
+                  set +u
+                  [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                  [[ "\$CERR" ]] && rm -f "\$CERR" || true
                   exit \$exit_status
                 }
